April 2024 commit including django-cms

+++ rtt/models.py
@@ -11,16 +11,16 @@ class NaPTANRailReferences(models.Model):
     name = models.CharField(max_length=100, blank=True, null=True)
     namelang = models.CharField(max_length=2, blank=True, null=True)
     gridtype = models.CharField(max_length=1, blank=True, null=True)
-    easting = models.PositiveIntegerField(blank=True, null=True)
-    northing = models.PositiveIntegerField(blank=True, null=True)
+    easting = models.FloatField(blank=True, null=True)
+    northing = models.FloatField(blank=True, null=True)
     creationdatetime = models.DateTimeField
     modificationdatetime = models.DateTimeField
     revisionnumber = models.SmallIntegerField(blank=True, null=True)
     modification = models.CharField(max_length=3, blank=True, null=True)
 
     class Meta:
-        verbose_name = 'Naptan Rail Reference'
-        verbose_name_plural = 'Naptan Rail References'
+        verbose_name = "Naptan Rail Reference"
+        verbose_name_plural = "Naptan Rail References"
 
     def __str__(self):
         return self.name
